get encryption working in the browser
